+++ example/usermanagement/serve_aiohttp_marshmallow.py
@@ -4,14 +4,15 @@ from datetime import datetime
 import json
 import time
 
+from aiohttp import web
 import bottle
 
-from example.usermanagement.schema import AboutSchema
-from example.usermanagement.schema import NoContentSchema
-from example.usermanagement.schema import UserAvatarSchema
-from example.usermanagement.schema import UserDigestSchema
-from example.usermanagement.schema import UserIdPathSchema
-from example.usermanagement.schema import UserSchema
+from example.usermanagement.schema_marshmallow import AboutSchema
+from example.usermanagement.schema_marshmallow import NoContentSchema
+from example.usermanagement.schema_marshmallow import UserAvatarSchema
+from example.usermanagement.schema_marshmallow import UserDigestSchema
+from example.usermanagement.schema_marshmallow import UserIdPathSchema
+from example.usermanagement.schema_marshmallow import UserSchema
 from example.usermanagement.userlib import User
 from example.usermanagement.userlib import UserAvatarNotFound
 from example.usermanagement.userlib import UserLib
@@ -21,7 +22,7 @@ from hapic import MarshmallowProcessor
 from hapic.data import HapicData
 from hapic.data import HapicFile
 from hapic.error.marshmallow import MarshmallowDefaultErrorBuilder
-from hapic.ext.bottle import BottleContext
+from hapic.ext.aiohttp.context import AiohttpContext
 
 try:  # Python 3.5+
     from http import HTTPStatus
@@ -29,13 +30,15 @@ except ImportError:
     from http import client as HTTPStatus
 
 
-hapic = Hapic()
+
+
+hapic = Hapic(async_=True)
 hapic.set_processor_class(MarshmallowProcessor)
 
-class BottleController(object):
+class AiohttpController(object):
     @hapic.with_api_doc()
     @hapic.output_body(AboutSchema())
-    def about(self):
+    async def about(self):
         """
         This endpoint allow to check that the API is running. This description
         is generated from the docstring of the method.
@@ -47,7 +50,7 @@ class BottleController(object):
 
     @hapic.with_api_doc()
     @hapic.output_body(UserDigestSchema(many=True))
-    def get_users(self):
+    async def get_users(self):
         """
         Obtain users list.
         """
@@ -57,7 +60,7 @@ class BottleController(object):
     @hapic.handle_exception(UserNotFound, HTTPStatus.NOT_FOUND)
     @hapic.input_path(UserIdPathSchema())
     @hapic.output_body(UserSchema())
-    def get_user(self, id, hapic_data: HapicData):
+    async def get_user(self, id, hapic_data: HapicData):
         """
         Return a user taken from the list or return a 404
         """
@@ -68,7 +71,7 @@ class BottleController(object):
     # TODO - G.M - 2017-12-5 - Support exclude, only ?
     @hapic.input_body(UserSchema(exclude=('id',)))
     @hapic.output_body(UserSchema())
-    def add_user(self, hapic_data: HapicData):
+    async def add_user(self, hapic_data: HapicData):
         """
         Add a user to the list
         """
@@ -80,7 +83,7 @@ class BottleController(object):
     @hapic.handle_exception(UserNotFound, HTTPStatus.NOT_FOUND)
     @hapic.output_body(NoContentSchema(), default_http_code=204)
     @hapic.input_path(UserIdPathSchema())
-    def del_user(self, id, hapic_data: HapicData):
+    async def del_user(self, id, hapic_data: HapicData):
         UserLib().del_user(int(hapic_data.path['id']))
         return NoContentSchema()
 
@@ -89,7 +92,7 @@ class BottleController(object):
     @hapic.handle_exception(UserAvatarNotFound, HTTPStatus.NOT_FOUND)
     @hapic.input_path(UserIdPathSchema())
     @hapic.output_file(['image/png'])
-    def get_user_avatar(self, id, hapic_data: HapicData):
+    async def get_user_avatar(self, id, hapic_data: HapicData):
         return HapicFile(
             file_path=UserLib().get_user_avatar_path(user_id=(int(hapic_data.path['id'])))
         )
@@ -100,45 +103,49 @@ class BottleController(object):
     @hapic.input_path(UserIdPathSchema())
     @hapic.input_files(UserAvatarSchema())
     @hapic.output_body(NoContentSchema(), default_http_code=204)
-    def update_user_avatar(self, id, hapic_data: HapicData):
+    async def update_user_avatar(self, id, hapic_data: HapicData):
         UserLib().update_user_avatar(
             user_id=int(hapic_data.path['id']),
             avatar=hapic_data.files['avatar'],
         )
 
-    def bind(self, app:bottle.Bottle):
-        app.route('/about', callback=self.about)
-        app.route('/users', callback=self.get_users)
-        app.route('/users/<id>', callback=self.get_user)
-        app.route('/users', callback=self.add_user,  method='POST')
-        app.route('/users/<id>', callback=self.del_user, method='DELETE')
-        app.route('/users/<id>/avatar', callback=self.get_user_avatar)
-        app.route('/users/<id>/avatar', callback=self.update_user_avatar, method='PUT')
+    def bind(self, app: web.Application):
+        app.add_routes([
+            web.get('/about', self.about),
+            web.get('/users', self.get_users),
+            web.get(r'/users/{id}', self.get_user),
+            web.post('/users/', self.add_user),
+            web.delete('/users/{id}', self.del_user),
+            web.get('/users/{id}/avatar', self.get_user_avatar),
+            web.put('/users/{id}/avatar', self.update_user_avatar)
+        ])
 
 
 if __name__ == "__main__":
-    app = bottle.Bottle()
-    controllers = BottleController()
+    app = web.Application()
+    controllers = AiohttpController()
     controllers.bind(app)
-    hapic.set_context(BottleContext(app, default_error_builder=MarshmallowDefaultErrorBuilder()))
-
-    print('')
-    print('')
-    print('GENERATING OPENAPI DOCUMENTATION')
-
-
+    hapic.set_context(
+        AiohttpContext(
+            app,
+            default_error_builder=MarshmallowDefaultErrorBuilder(),
+        ),
+    )
     doc_title = 'Demo API documentation'
     doc_description = 'This documentation has been generated from ' \
                        'code. You can see it using swagger: ' \
                        'http://editor2.swagger.io/'
     hapic.add_documentation_view('/doc/', doc_title, doc_description)
+    print('')
+    print('')
+    print('GENERATING OPENAPI DOCUMENTATION')
     openapi_file_name = 'api-documentation.json'
     with open(openapi_file_name, 'w') as openapi_file_handle:
         openapi_file_handle.write(
             json.dumps(
                 hapic.generate_doc(
                     title=doc_title,
-                    description=doc_description,
+                    description=doc_description
                 )
             )
         )
@@ -148,7 +155,7 @@ if __name__ == "__main__":
 
     print('')
     print('')
-    print('RUNNING BOTTLE SERVER NOW')
+    print('RUNNING AIOHTTP SERVER NOW')
     print('DOCUMENTATION AVAILABLE AT /doc/')
     # Run app
-    app.run(host='127.0.0.1', port=8081, debug=True)
+    web.run_app(app=app, host='127.0.0.1', port=8084)
